refactor: extract entity reset services
